implement native video on article page
let homeB be homeA

add caption for youtube on article page

fixed layout issue on home page
