+++ src/router/index.js
@@ -9,7 +9,7 @@ Vue.use(Meta)
 import NotFound from '../views/404.vue'
 import Article from '../views/Article.vue'
 import Home from '../views/Home.vue'
-import HomeB from '../views/HomeB.vue'
+// import HomeB from '../views/HomeB.vue'
 import List from '../views/List.vue'
 import Questionnaire from '../views/Questionnaire.vue'
 import Search from '../views/Search.vue'
@@ -29,7 +29,7 @@ const router = new Router({
         component: Article
       } ]
     },
-    { path: '/homeb', component: HomeB },
+    // { path: '/homeb', component: HomeB },
     { path: '/q/:questionnaireId/:resultId?', component: Questionnaire },
     { path: '/search/:keyword', component: Search },
     { path: '/section/:title', component: List },
